[PD] moves Extrude code to FeatureExtrude
- since the Extrude code is only used by Pocket/pad, it should better in FeatureExtrude
