+++ src/Mod/PartDesign/App/FeatureExtrude.h
@@ -30,6 +30,10 @@
 #include <Base/Tools.h>
 #include "FeatureSketchBased.h"
 
+class gp_Dir;
+class TopoDS_Face;
+class TopoDS_Shape;
+
 namespace PartDesign
 {
 
@@ -62,6 +66,54 @@ public:
 
 protected:
     Base::Vector3d computeDirection(const Base::Vector3d& sketchVector);
+
+    /**
+      * Generates an extrusion of the input sketchshape and stores it in the given &prism
+      */
+    void Extrude(TopoDS_Shape& prism,
+        const TopoDS_Shape& sketchshape,
+        const std::string& method,
+        const gp_Dir& direction,
+        const double L,
+        const double L2,
+        const double angle,
+        const double angle2,
+        const bool midplane,
+        const bool reversed);
+
+    // See BRepFeat_MakePrism
+    enum PrismMode {
+        CutFromBase = 0,
+        FuseWithBase = 1,
+        None = 2
+    };
+
+    /**
+      * Generates an extrusion of the input profileshape
+      * It will be a stand-alone solid created with BRepFeat_MakePrism
+      */
+    static void Extrude(TopoDS_Shape& prism,
+        const std::string& method,
+        const TopoDS_Shape& baseshape,
+        const TopoDS_Shape& profileshape,
+        const TopoDS_Face& sketchface,
+        const TopoDS_Face& uptoface,
+        const gp_Dir& direction,
+        PrismMode Mode,
+        Standard_Boolean Modify);
+
+    /**
+      * Generates a tapered prism of the input sketchshape and stores it in the given &prism
+      */
+    void generateTaperedPrism(TopoDS_Shape& prism,
+        const TopoDS_Shape& sketchshape,
+        const std::string& method,
+        const gp_Dir& direction,
+        const double L,
+        const double L2,
+        const double angle,
+        const double angle2,
+        const bool midplane);
 };
 
 } //namespace PartDesign
